moved e2e tests out of integration folder
