+++ .github/workflows/main.yml
@@ -135,7 +135,7 @@ jobs:
       - name: Run Cypress end to end tests
         uses: cypress-io/github-action@v2
         with:
-          spec: cypress/integration/endToEnd/*
+          spec: cypress/endToEnd/*
           browser: chrome
           start: npm run start
           build: npm run build
