+++ README.md
@@ -85,7 +85,7 @@ The UI tests are a part of our CI/CD pipeline.
 - Cypress needs the `SESSION_TOKEN` environment variable to be set
   - This value is set in env.local
   - The value is in 1Password - the token has been set to expire in 2031
-- `npm run cypress:open` or `npm run cypress:watch` -- Run Cypress tests in the `cypress/integration/endToEnd` directory with the Test Runner
+- `npm run cypress:open` or `npm run cypress:watch` -- Run Cypress tests in the `cypress/endToEnd` directory with the Test Runner
 - `npm run cypress:e2e` -- Run Cypress tests in the `cypress/intergration/endToEnd` directory in the command line
 
 The end to end tests are also a part of our CI/CD pipeline
