+++ cypress/endToEnd/end-to-end.spec.ts
@@ -1,30 +1,30 @@
 import {
   AccountPageURLs,
   CreateRegistrationPageURLs,
-} from "../../../src/lib/urls";
-import { sentenceCase } from "../../../src/lib/writingStyle";
+} from "../../src/lib/urls";
+import { sentenceCase } from "../../src/lib/writingStyle";
 import {
   testBeaconAndOwnerData,
   testLandUseData,
-} from "../../common/happy-path-test-data.spec";
+} from "../common/happy-path-test-data.spec";
 import {
   givenIHaveFilledInUpdateAccountDetailsPage,
   iCanSeeMyAccountDetails,
-} from "../../common/i-can-enter-account-details.spec";
-import { givenIHaveEnteredMyBeaconDetails } from "../../common/i-can-enter-beacon-information.spec";
+} from "../common/i-can-enter-account-details.spec";
+import { givenIHaveEnteredMyBeaconDetails } from "../common/i-can-enter-beacon-information.spec";
 import {
   givenIHaveEnteredMyAddressDetails,
   givenIHaveEnteredMyEmergencyContactDetails,
   givenIHaveEnteredMyPersonalDetails,
-} from "../../common/i-can-enter-owner-information.spec";
-import { andIHaveNoFurtherUses } from "../../common/i-can-enter-use-information/generic.spec";
-import { givenIHaveEnteredMyLandUse } from "../../common/i-can-enter-use-information/land.spec";
+} from "../common/i-can-enter-owner-information.spec";
+import { andIHaveNoFurtherUses } from "../common/i-can-enter-use-information/generic.spec";
+import { givenIHaveEnteredMyLandUse } from "../common/i-can-enter-use-information/land.spec";
 import {
   givenIHaveACookieSetAndHaveSignedInIVisit,
   givenIHaveClicked,
   givenIHaveWaitedForBeaconsApi,
   thenTheUrlShouldContain,
-} from "../../common/selectors-and-assertions.spec";
+} from "../common/selectors-and-assertions.spec";
 
 describe("As user with an account", () => {
   const iCanSeeTheBeaconListWithMyInformation = (): void => {
