+++ cypress/endToEnd/i-can-update-a-registration-with-many-uses.spec.ts
@@ -1,38 +1,38 @@
-import { BeaconUse } from "../../../src/entities/BeaconUse";
-import { DraftRegistration } from "../../../src/entities/DraftRegistration";
-import { Registration } from "../../../src/entities/Registration";
+import { BeaconUse } from "../../src/entities/BeaconUse";
+import { DraftRegistration } from "../../src/entities/DraftRegistration";
+import { Registration } from "../../src/entities/Registration";
 import {
   Activity,
   Environment,
   Purpose,
-} from "../../../src/lib/deprecatedRegistration/types";
-import { AccountPageURLs } from "../../../src/lib/urls";
-import { Actions } from "../../../src/lib/URLs/Actions";
-import { Pages } from "../../../src/lib/URLs/Pages";
-import { Resources } from "../../../src/lib/URLs/Resources";
-import { UsePages } from "../../../src/lib/URLs/UsePages";
-import { makeEnumValueUserFriendly } from "../../../src/lib/writingStyle";
-import { iAmPromptedToConfirm } from "../../common/i-am-prompted-to-confirm.spec";
+} from "../../src/lib/deprecatedRegistration/types";
+import { AccountPageURLs } from "../../src/lib/urls";
+import { Actions } from "../../src/lib/URLs/Actions";
+import { Pages } from "../../src/lib/URLs/Pages";
+import { Resources } from "../../src/lib/URLs/Resources";
+import { UsePages } from "../../src/lib/URLs/UsePages";
+import { makeEnumValueUserFriendly } from "../../src/lib/writingStyle";
+import { iAmPromptedToConfirm } from "../common/i-am-prompted-to-confirm.spec";
 import {
   andIHaveEnteredMyAviationUse,
   givenIHaveEnteredInformationAboutMyAircraft,
   givenIHaveEnteredMyAircraftCommunicationDetails,
   iCanSeeMyAviationUse,
-} from "../../common/i-can-enter-use-information/aviation.spec";
-import { whenIHaveAnotherUse } from "../../common/i-can-enter-use-information/generic.spec";
+} from "../common/i-can-enter-use-information/aviation.spec";
+import { whenIHaveAnotherUse } from "../common/i-can-enter-use-information/generic.spec";
 import {
   andIHaveEnteredMyLandUse,
   givenIHaveEnteredMyLandCommunicationDetails,
   iCanSeeMyLandUse,
-} from "../../common/i-can-enter-use-information/land.spec";
+} from "../common/i-can-enter-use-information/land.spec";
 import {
   givenIHaveEnteredInformationAboutMyVessel,
   givenIHaveEnteredMyVesselCommunicationDetails,
-} from "../../common/i-can-enter-use-information/maritime.spec";
+} from "../common/i-can-enter-use-information/maritime.spec";
 import {
   iHavePreviouslyRegisteredABeacon,
   randomUkEncodedHexId,
-} from "../../common/i-have-previously-registered-a-beacon.spec";
+} from "../common/i-have-previously-registered-a-beacon.spec";
 import {
   andIClickContinue,
   givenIHaveSelected,
@@ -45,10 +45,10 @@ import {
   whenIClickContinue,
   whenIClickTheButtonContaining,
   whenIHaveVisited,
-} from "../../common/selectors-and-assertions.spec";
-import { theNumberOfUsesIs } from "../../common/there-are-n-uses.spec";
-import { whenIGoToDeleteMy } from "../../common/when-i-go-to-delete-my.spec";
-import { singleBeaconRegistration } from "../../fixtures/singleBeaconRegistration";
+} from "../common/selectors-and-assertions.spec";
+import { theNumberOfUsesIs } from "../common/there-are-n-uses.spec";
+import { whenIGoToDeleteMy } from "../common/when-i-go-to-delete-my.spec";
+import { singleBeaconRegistration } from "../fixtures/singleBeaconRegistration";
 
 describe("As an account holder", () => {
   it("I can add many uses to one of my registrations", () => {
