+++ cypress/endToEnd/i-can-view-a-legacy-beacon.spec.ts
@@ -1,6 +1,6 @@
-import { iHavePreviouslyRegisteredALegacyBeacon } from "../../common/i-have-a-legacy-beacon.spec";
-import { givenIHaveSignedIn } from "../../common/selectors-and-assertions.spec";
-import { legacyBeaconRequest } from "../../fixtures/migration";
+import { iHavePreviouslyRegisteredALegacyBeacon } from "../common/i-have-a-legacy-beacon.spec";
+import { givenIHaveSignedIn } from "../common/selectors-and-assertions.spec";
+import { legacyBeaconRequest } from "../fixtures/migration";
 
 describe("As an account holder", () => {
   it.only("I can view a legacy beacon linked to my email", () => {
